refactor: 찜한 목록 폴더 정리
