+++ src/features/heart/api/deletePreAuctionHeart.ts
@@ -1,0 +1,7 @@
+import { API_END_POINT, httpClient } from '@/shared';
+
+export const deletePreAuctionHeart = async (preAuctionId: number) => {
+  await httpClient.post(`${API_END_POINT.PRE_AUCTION}/${preAuctionId}/likes`);
+
+  return;
+};
