+++ src/features/heart/ui/index.ts
@@ -1,0 +1,1 @@
+export { HeartList } from './HeartList';
