+++ src/pages/heart/ui/Heart.tsx
@@ -1,34 +1,9 @@
-import { Button, EmptyBoundary, ROUTES } from '@/shared';
-import { useDeletePreAuctionHeart, useGetPreAuctionHeartList } from '@/components/heart/queries';
-
-import AuctionItem from '@/entities/auction/ui/AuctionItem';
-import type { IPreAuctionItem } from '@/@types/AuctionItem';
-import { useNavigate } from 'react-router-dom';
+import { HeartList } from "@/features/heart";
 
 export const Heart = () => {
-  const navigate = useNavigate();
-  const { preAuctionHeartList } = useGetPreAuctionHeartList();
-  const { mutate } = useDeletePreAuctionHeart();
-
-  const handleDelete = (id: number) => mutate(id);
-
   return (
-    <EmptyBoundary length={preAuctionHeartList.length} name='heart'>
-      <ul className='grid items-center justify-between grid-cols-2 gap-4'>
-        {preAuctionHeartList.map((el: IPreAuctionItem) => (
-          <li key={el.productId} onClick={() => navigate(ROUTES.PRE_AUCTION.getItemRoute(el.productId))} className='cursor-pointer'>
-            <AuctionItem axis='column' label='내가 찜 한 사전 경매 상품'>
-              <AuctionItem.Image src={el.imageUrl} />
-              <AuctionItem.Main name={el.productName} count={el.likeCount} price={el.minPrice} kind='pre-register' />
-              <AuctionItem.Button>
-                <Button onClick={() => handleDelete(el.productId)} type='button' className='w-full'>
-                  찜 목록에서 제외
-                </Button>
-              </AuctionItem.Button>
-            </AuctionItem>
-          </li>
-        ))}
-      </ul>
-    </EmptyBoundary>
+    <ul className='grid items-center justify-between grid-cols-2 gap-4'>
+      <HeartList />
+    </ul>
   );
 };
